pyproject.toml
[project]
name = "eventman"
version = "0.0.1"
dependencies = [
    "django>=5.1,<6.0",
    "python-decouple>=3.8",
    "dj-database-url>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "django-allauth>=65.0.0",
    "django-debug-toolbar>=4.4.0",
    "django-htmx>=1.19.0",
    "django-crispy-forms>=2.3",
    "crispy-bootstrap5>=2024.2",
    "django-filter>=24.3",
    "django-model-utils>=5.0.0",
    "django-braces>=1.16.0",
    "pillow>=10.4.0",
    "whitenoise>=6.7.0",
    "gunicorn>=23.0.0",
    "upstash-redis>=0.15.0",
    "django-redis>=5.4.0",
    "python-dotenv>=1.0.1",
    "setuptools>=75.0.0",
    "cloudinary>=1.39.0",
    "django-cloudinary-storage>=0.0.12",
    "hammett>=0.1.0",
    "sslcommerz-lib>=1.0",
    "django-widget-tweaks>=1.5.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "factory-boy>=3.3.3",
    "faker>=37.11.0",
    "autoprefixer>=0.1.0",
]

[tool.uv]
package = false

[tool.hammett]
plugins = ["pytest_django", "pytest_mock"]
